+++ ChangeLog.md
@@ -1,8 +1,10 @@
-# Changelog for decimal64
+# Changelog for safe-decimal
 
 ## 0.2.0
 
 * Make rounding dependent on precision type (extra type argument to the `Round` type class)
+* Rename `RoundFloor` to `RoundDown` (with `Floor` as synonym)
+* Rename `Truncate` to `RoundToZero` (with `Truncate` as synonym)
 * Fix conversion `fromRational` to throw `Underflow` instead of rounding. Without this fix
   there was unexpected behavior on negative numbers.
 * Addition of `scaleUp` and `scaleUpBounded`